+++ cmsWidgets/breadcrumbs/BreadcrumbsCmsWidget.php
@@ -8,13 +8,7 @@
 
 namespace skeeks\cms\cmsWidgets\breadcrumbs;
 
-use skeeks\cms\base\Widget;
 use skeeks\cms\base\WidgetRenderable;
-use skeeks\cms\helpers\UrlHelper;
-use yii\helpers\ArrayHelper;
-use yii\helpers\Html;
-use yii\helpers\Json;
-
 /**
  * Class breadcrumbs
  * @package skeeks\cms\cmsWidgets\Breadcrumbs
@@ -24,7 +18,7 @@ class BreadcrumbsCmsWidget extends WidgetRenderable
     static public function descriptorConfig()
     {
         return array_merge(parent::descriptorConfig(), [
-            'name' => \Yii::t('skeeks/cms','Breadcrumbs')
+            'name' => \Yii::t('skeeks/cms', 'Breadcrumbs')
         ]);
     }
 }